services/gateway/requirements.txt
Django
djangorestframework
psycopg2-binary
Pillow
requests
gunicorn
python-dotenv
django-cors-headers
djangorestframework-simplejwt
psycopg[binary]